refactor code + separate compoenents
