+++ lib/utils/communities.ts
@@ -1,4 +1,4 @@
-import communitiesData from "../data/communities.json"
+import communitiesData from "../../public/bitcoin-topics/communities/communities.json"
 
 // Types
 export type WeeklyData = {
@@ -97,6 +97,16 @@ export type CohortStatus = "in-progress" | "starting-soon" | "completed"
 type BaseQuestion = {
     text: string
     relatedMaterial: string
+}
+
+export type Question = {
+    id: string
+    text: string
+    relatedMaterial?: string
+    assignments: Array<{
+        studentName: string
+        groupId: number
+    }>
 }
 
 // distribute questions among students
@@ -283,4 +293,130 @@ export const calculateCohortProgress = (community: Community): number => {
     const progress = Math.round((elapsed / totalDuration) * 100)
 
     return Math.min(Math.max(progress, 0), 100)
+}
+
+export const getNextCallDate = (weeklyCall: WeeklyData["weeklyCall"]) => {
+    const callDay = new Date(weeklyCall.date)
+    callDay.setHours(weeklyCall.time.hour, weeklyCall.time.minute, 0, 0)
+    return callDay
+}
+
+export const isCallActive = (weeklyCall: WeeklyData["weeklyCall"]) => {
+    const now = new Date()
+    const nextCall = getNextCallDate(weeklyCall)
+    const timeDiff = nextCall.getTime() - now.getTime()
+    const minutesUntilCall = timeDiff / (1000 * 60)
+    return minutesUntilCall <= 10 && minutesUntilCall >= -60
+}
+
+export const createCalendarEvent = (weeklyCall: WeeklyData["weeklyCall"]) => {
+    const timeZone = weeklyCall.time.timeZone
+    const nextCall = new Date(weeklyCall.date)
+    debugger
+    // Create a date in the specified timezone
+    const userTimeZone = Intl.DateTimeFormat().resolvedOptions().timeZone
+    const formatter = new Intl.DateTimeFormat("en-US", {
+        timeZone: timeZone,
+        year: "numeric",
+        month: "2-digit",
+        day: "2-digit",
+        hour: "2-digit",
+        minute: "2-digit",
+        hour12: false
+    })
+
+    // Set the time in the original timezone
+    const originalDate = new Date(weeklyCall.date)
+    originalDate.setHours(weeklyCall.time.hour, weeklyCall.time.minute, 0, 0)
+
+    // Format the date in the original timezone
+    const formattedDate = formatter.format(originalDate)
+    const [datePart, timePart] = formattedDate.split(", ")
+    const [month, day, year] = datePart.split("/")
+    const [hour, minute] = timePart.split(":")
+
+    // Create the start date in the correct timezone
+    nextCall.setFullYear(parseInt(year), parseInt(month) - 1, parseInt(day))
+    nextCall.setHours(parseInt(hour), parseInt(minute), 0, 0)
+
+    // Create end time (1 hour later)
+    const endCall = new Date(nextCall)
+    endCall.setHours(endCall.getHours() + 1)
+
+    const event = {
+        title: "Bitcoin Development Weekly Call",
+        description: weeklyCall.description,
+        startTime: nextCall
+            .toISOString()
+            .replace(/[-:]/g, "")
+            .replace(/\.\d{3}/, ""),
+        endTime: endCall
+            .toISOString()
+            .replace(/[-:]/g, "")
+            .replace(/\.\d{3}/, ""),
+        location: `Discord - ${weeklyCall.discordLink}`,
+        timezone: timeZone
+    }
+
+    const icsContent = `BEGIN:VCALENDAR
+VERSION:2.0
+PRODID:-//Decoding Bitcoin//Calendar//EN
+BEGIN:VEVENT
+SUMMARY:${event.title}
+DESCRIPTION:${event.description.replace(/\n/g, "\\n")}
+DTSTART;TZID=${event.timezone}:${event.startTime}
+DTEND;TZID=${event.timezone}:${event.endTime}
+LOCATION:${event.location}
+RRULE:FREQ=WEEKLY
+BEGIN:VALARM
+TRIGGER:-PT30M
+ACTION:DISPLAY
+DESCRIPTION:Reminder
+END:VALARM
+END:VEVENT
+END:VCALENDAR`
+
+    const blob = new Blob([icsContent], { type: "text/calendar;charset=utf-8" })
+    const link = document.createElement("a")
+    link.href = window.URL.createObjectURL(blob)
+    link.download = "bitcoin-development-call.ics"
+    document.body.appendChild(link)
+    link.click()
+    document.body.removeChild(link)
+}
+
+export const getTimeUntilCall = (weeklyCall: WeeklyData["weeklyCall"]) => {
+    const nextCall = getNextCallDate(weeklyCall)
+    const now = new Date()
+    const diff = nextCall.getTime() - now.getTime()
+
+    const days = Math.floor(diff / (1000 * 60 * 60 * 24))
+    const hours = Math.floor((diff % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60))
+    const minutes = Math.floor((diff % (1000 * 60 * 60)) / (1000 * 60))
+    const seconds = Math.floor((diff % (1000 * 60)) / 1000)
+
+    const pad = (n: number) => Math.abs(n).toString().padStart(2, "0")
+
+    return { days, hours, minutes, seconds, pad }
+}
+
+export const collectQuestions = (weekData: WeeklyData): Question[] => {
+    if (!weekData || !weekData.questions) return []
+
+    const questionMap = new Map<string, Question>()
+
+    weekData.questions.forEach((question) => {
+        const key = question.text
+        if (!questionMap.has(key)) {
+            questionMap.set(key, {
+                id: question.id,
+                text: question.text,
+                relatedMaterial: question.relatedMaterial,
+                assignments: []
+            })
+        }
+        questionMap.get(key)?.assignments.push(question.assignedTo)
+    })
+
+    return Array.from(questionMap.values())
 }
